+++ 1.20/src/main/java/io/github/axolotlclient/modules/hypixel/bedwars/BedwarsGame.java
@@ -400,8 +400,10 @@ public class BedwarsGame {
 			while (this.seconds % 60 != target) {
 				updateClock();
 			}
-			topBarText = Text.literal(calculateTopBarText());
-			bottomBarText = Text.literal(calculateBottomBarText());
+			if (me != null) {
+				topBarText = Text.literal(calculateTopBarText());
+				bottomBarText = Text.literal(calculateBottomBarText());
+			}
 		}
 	}